requirements.txt
python-telegram-bot
flask
uvicorn
asgiref
python-dotenv
requests